feat: implement robust startup configuration validation layer
- Add ConfigValidator to enforce production safety standards.
- Validate SYMBOL, TIMEFRAME, and risk limits from RISK_LIMITS.md.
- Block launch for critical failures (accuracy < 0.50, win rate < 0.45, ECE > 0.25).
- Detect placeholder secrets in REDIS_URL and MT5_SERVER.
- Fix linting and test regressions.
- Update DEPLOYMENT_GUIDE.md and CHANGELOG.md.

+++ src/research/hyperopt_walkforward.py
@@ -247,7 +247,7 @@ class WalkForwardOptimizer:
 
                     # Ensure type safety
                     if is_int:
-                        new_val = int(round(new_val))
+                        new_val = round(new_val)
 
                     # If no change occurred (e.g. value was small and delta was 0), skip
                     if new_val == original_val:
